+++ nf_core/lint.py
@@ -201,6 +201,8 @@ class PipelineLint(object):
             'check_actions_branch_protection',
             'check_actions_ci',
             'check_actions_lint',
+            'check_actions_awstest',
+            'check_actions_awsfulltest',
             'check_readme',
             'check_conda_env_yaml',
             'check_conda_dockerfile',
@@ -236,16 +238,18 @@ class PipelineLint(object):
             'CHANGELOG.md',
             'docs/README.md',
             'docs/output.md',
-            'docs/usage.md'
+            'docs/usage.md',
+            '.github/workflows/branch.yml',
+            '.github/workflows/ci.yml',
+            '.github/workflows/linting.yml'
 
         Files that *should* be present::
 
             'main.nf',
             'environment.yml',
             'conf/base.config',
-            '.github/workflows/branch.yml',
-            '.github/workflows/ci.yml',
-            '.github/workfows/linting.yml'
+            '.github/workflows/awstest.yml',
+            '.github/workflows/awsfulltest.yml'
 
         Files that *must not* be present::
 
@@ -277,7 +281,9 @@ class PipelineLint(object):
         files_warn = [
             ['main.nf'],
             ['environment.yml'],
-            [os.path.join('conf','base.config')]
+            [os.path.join('conf','base.config')],
+            [os.path.join('.github', 'workflows','awstest.yml')],
+            [os.path.join('.github', 'workflows', 'awsfulltest.yml')]
         ]
 
         # List of strings. Dails / warns if any of the strings exist.
@@ -673,6 +679,66 @@ class PipelineLint(object):
                 self.failed.append((5, "Continuous integration must run nf-core lint Tests: '{}'".format(fn)))
             else:
                 self.passed.append((5, "Continuous integration runs nf-core lint Tests: '{}'".format(fn)))
+
+    def check_actions_awstest(self):
+        """Checks the GitHub Actions awstest is valid.
+
+        Makes sure it is triggered only on ``push`` to ``master``. 
+        """
+        fn = os.path.join(self.path, '.github', 'workflows', 'awstest.yml')
+        if os.path.isfile(fn):
+            with open(fn, 'r') as fh:
+                wf = yaml.safe_load(fh)
+            
+            # Check that the action is only turned on for push
+            try:
+                assert('push' in wf[True])
+                assert('pull_request' not in wf[True])
+            except (AssertionError, KeyError, TypeError):
+                self.failed.append((5, "GitHub Actions AWS test should be triggered on push and not PRs: '{}'".format(fn)))
+            else:
+                self.passed.append((5, "GitHub Actions AWS test is triggered on push and not PRs: '{}'".format(fn)))
+
+            # Check that the action is only turned on for push to master
+            try:
+                assert('master' in wf[True]['push']['branches'])
+                assert('dev' not in wf[True]['push']['branches'])
+            except (AssertionError, KeyError, TypeError):
+                self.failed.append((5, "GitHub Actions AWS test should be triggered only on push to master: '{}'".format(fn)))
+            else:
+                self.passed.append((5, "GitHub Actions AWS test is triggered only on push to master: '{}'".format(fn)))
+            
+    def check_actions_awsfulltest(self):
+        """Checks the GitHub Actions awsfulltest is valid.
+
+        Makes sure it is triggered only on ``release``.
+        """
+        fn = os.path.join(self.path, '.github', 'workflows', 'awsfulltest.yml')
+        if os.path.isfile(fn):
+            with open(fn, 'r') as fh:
+                wf = yaml.safe_load(fh)
+
+            aws_profile = '-profile test '
+
+            # Check that the action is only turned on for published releases
+            try:
+                assert('release' in wf[True])
+                assert('published' in wf[True]['release']['types'])
+                assert('push' not in wf[True])
+                assert('pull_request' not in wf[True])
+            except (AssertionError, KeyError, TypeError):
+                self.failed.append((5, "GitHub Actions AWS full test should be triggered only on published release: '{}'".format(fn)))
+            else:
+                self.passed.append((5, "GitHub Actions AWS full test is triggered only on published release: '{}'".format(fn)))
+
+            # Warn if `-profile test` is still unchanged
+            try:
+                steps = wf['jobs']['run-awstest']['steps']
+                assert(any([aws_profile in step['run'] for step in steps if 'run' in step.keys()]))
+            except (AssertionError, KeyError, TypeError):
+                self.passed.append((5, "GitHub Actions AWS full test should test full datasets: '{}'".format(fn)))
+            else:
+                self.warned.append((5, "GitHub Actions AWS full test should test full datasets: '{}'".format(fn)))
 
     def check_readme(self):
         """Checks the repository README file for errors.