+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Update `output.md` and add in 'Pipeline information' section describing standard NF and pipeline reporting.
 * Build Docker image using GitHub Actions, then push to Docker Hub (instead of building on Docker Hub)
 * New Slack channel badge in pipeline readme
+* Add AWS CI tests and full tests GitHub Actions workflows
 
 ### Linting
 
@@ -25,6 +26,7 @@
 * Fix failure when providing version along with build id for Conda packages
 * New `--json` and `--markdown` options to print lint results to JSON / markdown files
 * Linting code now automatically posts warning / failing results to GitHub PRs as a comment if it can
+* Added AWS GitHub Actions workflows linting
 
 ### Other